% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agencies.R
\name{get_agencies}
\alias{get_agencies}
\title{get_agencies}
\usage{
get_agencies(jurisdictionID = NULL, keyword = NULL, verbose = 0)
}
\arguments{
\item{jurisdictionID}{ID for the jurisdiction}

\item{keyword}{search for keyword in agency name}

\item{verbose}{prints out the url of the API call, default value of 0}
}
\value{
Returns pandas dataframe with the metadata
}
\description{
Get metadata for all agencies of a specific jurisdiction
}
\examples{
get_agencies(jurisdictionID=38, keyword="test")
}
