% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_document_values.R
\name{get_document_values}
\alias{get_document_values}
\title{get_document_values}
\usage{
get_document_values(...)
}
\arguments{
\item{jurisdiction}{ID for the jurisdiction}

\item{year}{Year(s) of data}

\item{documentType}{ID for type of document, default value of 1}

\item{summary}{Return summary instead of document level data (only one year of data is allowed for document level data), default value is TRUE}

\item{dateIsRange}{Indicating whether the time parameter is range or should be treated as single data points, default value is TRUE}

\item{country}{Get values for all subjurisdictions, default value is FALSE}

\item{agency}{Agency ID, default value is NULL}

\item{label}{Industry code using the jurisdiction-specific coding system (returns all 3-digit industries by default), default value is NULL}

\item{filtered}{Exclude poorly-performing industry results (use of unfiltered results is NOT recommended), default value is TRUE}

\item{labellevel}{Level of NAICS industries to include, default value is 3}

\item{labelsource}{classification standard (NAICS, BEA, SOC), default value of "NAICS"}

\item{version}{Version ID for datasets with multiple versions (if no ID is given, returns most recent version), default value is NULL}

\item{verbose}{Print out the url of the API call (useful for debugging), default value is 0}
}
\value{
Returns pandas dataframe with the metadata
}
\description{
Get values for a specific jurisdiction and series at the document level
}
\examples{
get_document_values(series = 33, jurisdiction = 38, year = 2018, label = "111")
}
