% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_documents.R
\name{get_documents}
\alias{get_documents}
\title{get_documents}
\usage{
get_documents(
  documentID = NULL,
  jurisdictionID = NULL,
  date = NULL,
  documentType = 1,
  verbose = 0
)
}
\arguments{
\item{documentID}{ID of the specific document}

\item{jurisdictionID}{ID for the jurisdiction}

\item{date}{Year(s) of the documents}

\item{documentType}{ID for type of document, default value of 1}

\item{verbose}{prints out the url of the API call, default value of 0}
}
\value{
Returns pandas dataframe with the metadata
}
\description{
Get metadata for documents available in a specific jurisdiction or for a specific document ID
}
\examples{
get_documents(jurisdictionID = 38, date = 2018)
}
