% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_agencies.R
\name{list_agencies}
\alias{list_agencies}
\title{list_agencies}
\usage{
list_agencies(jurisdictionID = NULL, keyword = NULL, reverse = FALSE)
}
\arguments{
\item{jurisdictionID}{ID for the jurisdiction}

\item{keyword}{search for keyword in agency name}

\item{reverse}{reverses the key-value mapping, default value of FALSE}
}
\value{
Returns dictionary containing names of agencies and their associated IDs
}
\description{
Fetches the names of the agencies and their associated IDs
}
\examples{
list_agencies(jurisdictionID=38, reverse=TRUE)
}
