% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dates.R
\name{list_dates}
\alias{list_dates}
\title{list_dates}
\usage{
list_dates(jurisdictionID, documentType = NULL)
}
\arguments{
\item{jurisdictionID}{ID for the jurisdiction}

\item{documentType}{ID for type of document}
}
\value{
Returns list of dates available for the jurisdiction
}
\description{
Fetches the dates available for the jurisdiction
}
\examples{
list_dates(jurisdictionID = 38)
}
