% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_industries.R
\name{list_industries}
\alias{list_industries}
\title{list_industries}
\usage{
list_industries(
  keyword = NULL,
  labellevel = 3,
  labelsource = "NAICS",
  onlyID = FALSE,
  reverse = FALSE
)
}
\arguments{
\item{keyword}{search for keyword in agency name}

\item{labellevel}{NAICS level (2 to 6-digit), default value of 3}

\item{labelsource}{classification standard (NAICS, BEA, SOC), default value of "NAICS"}

\item{onlyID}{uses the NAICS code instead of name as key of dictionary, default value of FALSE}

\item{reverse}{reverses the key-value mapping, default value of FALSE}
}
\value{
Returns dictionary containing names of industries and associated IDs
}
\description{
Fetches the names of industries and associated IDs
}
\examples{
list_industries(onlyID=TRUE, reverse=FALSE)
}
